% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fitstats.R
\name{boot_fitstats}
\alias{boot_fitstats}
\title{Bootstrap Goodness-of-fit Statistics}
\usage{
boot_fitstats(
  n_sites,
  n_samps,
  lambda,
  sigma,
  det_prob,
  sampling_method = "distance",
  W = 20,
  nsim = 10,
  parallel = FALSE
)
}
\arguments{
\item{n_sites}{Number of sites to simulate}

\item{n_samps}{Number of samples (replicates) per site to simulate}

\item{lambda}{mean number of organisms at each site (1 draw per site)}

\item{sigma}{SD of half-normal detection function}

\item{det_prob}{mean detection probability (across distances from 0 to W)}

\item{sampling_method}{either "pointcount' or 'distance'}

\item{W}{Transect half-width}

\item{nsim}{number of simulations
# progress_bar = FALSE,}

\item{parallel}{Logical. Whether or not to use parallel processing for the simulations. Defaults to FALSE.}
}
\value{
data.frame with 3 columns (1 for each fit statistic) and nsim rows.
}
\description{
This function bootstraps p-values for the test statistics calculated with fitstats_optim().
It uses the sim_data() function in this package to simulate unbiased datasets.
}
