% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_det_prob.R
\name{calc_det_prob}
\alias{calc_det_prob}
\title{Calculate Mean Detection Probability}
\usage{
calc_det_prob(W, sigma)
}
\arguments{
\item{W}{transect half-width (in meters)}

\item{sigma}{Known SD of normal distribution describing the detection function (in meters)}
}
\value{
number between 0 and 1
}
\description{
This uses sigma (the SD of a normal distribution) and W (the transect half-width)
to calculate the mean detection probability. Only supports half-normal 
detection function at this point.
}
