% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sigma.R
\name{find_sigma}
\alias{find_sigma}
\title{Iteratively find sigma for half-normal detection function}
\usage{
find_sigma(start = 0.1, W = 20, p_true = 0.42)
}
\arguments{
\item{start}{starting guess (default is fine)}

\item{W}{known transect half-width}

\item{p_true}{known detection probability}
}
\value{
sd of normal distribution to the nearest 0.01
}
\description{
Because I'm too lazy to work out the math for calculating sigma (SD) precisely,
I created this function to iteratively find sigma (SD) for a half-normal 
detection function. It assumes a half-normal detection function with SD = sigma.
}
\examples{
find_sigma()

}
