% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitstats.R
\name{fitstats}
\alias{fitstats}
\title{Goodness-of-fit Statistics}
\usage{
fitstats(fm)
}
\arguments{
\item{fm}{a model fit in the unmarked package}
}
\value{
vector of 3 fit statistics
}
\description{
I copied this function from https://github.com/rbchan/unmarked/blob/master/R/boot.R
It's a function returning three goodness-of-fit statistics: sum-of-squares, Chi-squared, and the Freeman-Tukey test-statistic.
}
