% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitstats_optim.R
\name{fitstats_optim}
\alias{fitstats_optim}
\title{Goodness-of-fit Statistics}
\usage{
fitstats_optim(optim_out, n_obs, y_obs = NA, W, sampling_method)
}
\arguments{
\item{n_obs}{vector of observations at each site & replicate}

\item{y_obs}{matrix of the number of observations in each distance category (from formatted unmarked data). Only used if sampling_method = "distance"}

\item{W}{transect half-width (in meters)}

\item{sampling_method}{either "pointcount" or "distance"}

\item{out}{fit from optim}
}
\value{
vector of 3 fit statistics
}
\description{
I based this function on "fitstats()" from https://github.com/rbchan/unmarked/blob/master/R/boot.R
It's a function returning three goodness-of-fit statistics - sum-of-squares, 
Chi-squared, and the Freeman-Tukey test-statistic - calculated using my optim
estimates rather than a model fit in "unmarked".
}
