% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_N_distribution.R
\name{get_N_distribution}
\alias{get_N_distribution}
\title{Approximate Posterior Distribution of N}
\usage{
get_N_distribution(
  lambda_est,
  n_obs,
  K = NA,
  p_est = NA,
  sampling_method = "pointcount",
  if_distance__sigma_est = NA,
  if_distance__W = NA,
  if_distance__y_list = NA
)
}
\arguments{
\item{lambda_est}{estimated lambda value}

\item{n_obs}{number observed}

\item{K}{maximum N possible (typically obtained from n_obs)}

\item{p_est}{detection probability. Given if sampling_method == 'pointcount'. Calculated from if_distance__sigma_est if sampling_method == 'distance'}

\item{sampling_method}{either 'pointcount' or 'distance'}

\item{if_distance__sigma_est}{Estimated detection parameter, sigma. Only needed if sampling_method == distance}

\item{if_distance__W}{Transect half-width. Only needed if sampling_method == distance}

\item{if_distance__y_list}{array of distance data with row = site, col = repl, 3rd dim = observations. Only needed if sampling_method == distance}
}
\value{
PDF of possible counts. Formatted as an array of probabilities with row = site, col = repl, 3rd dim = observations.
}
\description{
Function for getting an estimate of the posterior distribution of N from my 
optim models. This is from the unmarked::ranef() function: https://github.com/rbchan/unmarked/blob/master/R/ranef.R
and is needed for calculating Observation rq residuals.
}
