% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll}
\alias{nll}
\title{Negative Log Likelihood Function for Estimating Sigma and Lambda}
\usage{
nll(par, K, n_dat, W = 20, y_dat, distance_sampling = TRUE)
}
\arguments{
\item{par}{parameters to be guessed (in this case c(log(sigma), log(lambda))). The parameters are logged so that they're always positive.}

\item{K}{The maximum possible true abundance, N. I've been using round(lambda * 10) as a default.}

\item{n_dat}{matrix of observations at all sites (rows) and samples/replicates (columns)}

\item{W}{Transect half-width (meters). If not set, it uses the maximum observed distance.}

\item{y_dat}{vector of distances (meters) for all observations}

\item{distance_sampling}{logical. Whether or not to incorporate distance data into likelihood.}
}
\value{
total negative log likelihood of the data given the parameters.
}
\description{
function description
}
