% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normtest.R
\name{normtest}
\alias{normtest}
\title{Test for Normality}
\usage{
normtest(x, name)
}
\arguments{
\item{x}{vector of values to be tested for normality}

\item{name}{name for saving the results iff there's an error}
}
\value{
If everything runs smoothly, this will return a vector with 3 values for each of the 3 tests of normality.
}
\description{
Run 3 different tests of normality, including shapiro.wilks (only if length(x) < 5000), 
AD test from package nortest, and lillie from nortest
}
\examples{
normtest(x = rnorm(100), 'name')

}
