% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_summary.R
\name{opt_summary}
\alias{opt_summary}
\title{Summary of Goodness-of-fit Statistics}
\usage{
opt_summary(fit_data, fit_sims)
}
\arguments{
\item{fit_data}{output from the fitstats_optim function}

\item{fit_sims}{output from the boot_fitstats function}
}
\value{
data.frame with fit_obs = fit statistic of observed data; "mean(fit_obs - fit_sim)" = mean bias in the fit statistics; "StdDev(fit_obs - fit_sim)" = SD of the bias in fit statistics; "Pr(fit_sim > fit_obs)" = p-value of the fit statistics
}
\description{
I based this function on https://github.com/rbchan/unmarked/blob/master/R/boot.R
It returns a summary of the 3 fit statistics calculated by comparing fit of 
observed data (fitstats_optim) to fit of datasets simulated with the model 
estimates (boot_fitstats).
}
