% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqRes0}
\alias{rqRes0}
\title{Helps calculate Randomized Quantile Residuals}
\usage{
rqRes0(a, b)
}
\arguments{
\item{a}{lower limit on uniform distribution = F(z-1) in Knape 2018}

\item{b}{upper limit on uniform distribution = F(z) in Knape 2018}
}
\value{
a single value
}
\description{
Used in the Generic function for Marginal, Site-sum, or Observation Randomized-quantile
residuals, as defined in Knape et al. 2018 and available for unmarked models
in the nmixgof package. This is a generic function used in the background of
other functions.
}
