% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResMar}
\alias{rqResMar}
\title{Marginal Randomized Quantile Residuals}
\usage{
rqResMar(umFit)
}
\arguments{
\item{umFit}{an unmarked model fit to point count data.}
}
\value{
a matrix of residuals with the same dimensions as the observations (rows = cites, columns = replicates)
}
\description{
Calculates Marginal Randomized-quantile residuals for a point count model fit with unmarked.
Use rqResMar_dist function for distance model.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
