% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResMar_optim}
\alias{rqResMar_optim}
\title{Marginal Randomized Quantile Residuals}
\usage{
rqResMar_optim(n_obs, lambda_est)
}
\arguments{
\item{n_obs}{a matrix of observed counts, with rows = sites and columns = replicates}

\item{lambda_est}{estimated lambda (mean abundance across sites) from fitted model}
}
\value{
a matrix of residuals with the same dimensions as the observations, n_obs.
}
\description{
Calculates Marginal Randomized-quantile residuals for a model fit with optim.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
