% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResObs_optim}
\alias{rqResObs_optim}
\title{Observation Randomized Quantile Residuals}
\usage{
rqResObs_optim(
  n_obs,
  K,
  lambda_est,
  sigma_est = NA,
  W = NA,
  p_est = NA,
  y_list = NA,
  sampling_method,
  show.warnings = FALSE
)
}
\arguments{
\item{n_obs}{a matrix of observed counts, with rows = sites and columns = replicates}

\item{K}{maximum possible true mean abundance, N}

\item{lambda_est}{estimated lambda (mean abundance across sites) from fitted model}

\item{sigma_est}{estimated sigma from fitted model (only needed if distance sampling)}

\item{W}{transect half-width. (only needed if distance sampling)}

\item{p_est}{estimated mean (across distances = 0 to W) detection probability (only needed if point counts)}

\item{y_list}{list of distance values with elements = sites and replicates elements within each element}

\item{sampling_method}{either 'pointcount' or 'distance'}
}
\value{
a matrix of residuals with the same dimensions as the observations, n_obs.
}
\description{
Calculates Observation Randomized-quantile residuals for a model fit with optim.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
