% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResSum}
\alias{rqResSum}
\title{Site-Sum Randomized Quantile Residuals}
\usage{
rqResSum(umFit, show.warnings = FALSE)
}
\arguments{
\item{umFit}{an unmarked model fit to point count data.}
}
\value{
a matrix of residuals with the same dimensions as the observations (rows = cites, columns = replicates)
}
\description{
Calculates Site-Sum Randomized-quantile residuals for a point count model fit with unmarked.
Use rqResSum_dist function for distance model.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
