% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResSum_dist}
\alias{rqResSum_dist}
\title{Site-Sum Randomized Quantile Residuals}
\usage{
rqResSum_dist(umFit, show.warnings = FALSE)
}
\arguments{
\item{umFit}{an unmarked model fit to point count data.}
}
\value{
a matrix of residuals with the same dimensions as the observations (rows = cites, columns = distance breaks)
}
\description{
Calculates Site-Sum Randomized-quantile residuals for a distance model fit with unmarked.
Note that I currently calculate these residuals by subtracting 1 from each
distance bin. That's pretty hacky and likely just plain wrong. Don't use
residuals from more than 1 distance bin. Just use residuals from a single bin.
It would be better to extend the observation rq residuals to a multinomial
distribution, and use the pbinsum function to get the possible N's (up to K)
across all distance breaks. But then I could just draw a random distance from
the normal distribution, take 1 individual out of that distance category (for
the F(z-1) calculation) and calculate the residuals with those 2 values &
distributions.
Use rqResSum function for point-count model.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
