% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqRes.R
\name{rqResSum_optim}
\alias{rqResSum_optim}
\title{Site-Sum Randomized Quantile Residuals}
\usage{
rqResSum_optim(n_obs, K, p_est, lambda_est, show.warnings = FALSE)
}
\arguments{
\item{n_obs}{a matrix of observed counts, with rows = sites and columns = replicates}

\item{K}{maximum possible true mean abundance, N}

\item{p_est}{estimated mean (across distances = 0 to W) detection probability}

\item{lambda_est}{estimated lambda (mean abundance across sites) from fitted model}
}
\value{
a matrix of residuals with the same dimensions as the observations, n_obs.
}
\description{
Calculates Site-Sum Randomized-quantile residuals for a model fit with optim.
As defined in Knape et al. 2018 and available for unmarked models in the nmixgof package.
}
