% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_N.R
\name{sim_data_N}
\alias{sim_data_N}
\title{Simulate biased (via over-dispersion in N) point count and distance sampling data (Scenario 2)}
\usage{
sim_data_N(
  n_sites = 50,
  n_samps = 6,
  lambda1 = 5,
  lambda2 = 5,
  det_prob = 0.42,
  sigma = NA,
  W = 20
)
}
\arguments{
\item{n_sites}{number of sites (transects)}

\item{n_samps}{number of samples (replicates) per site}

\item{lambda1}{partial mean abundance at every site (single draw per site that stays constant across samples)}

\item{lambda2}{partial mean abundance at every replicate (new draw for every sample at every site)}

\item{det_prob}{mean detection probability}

\item{sigma}{detection parameter (meters). Just leave this blank and it will be calculated from det_prob}

\item{W}{transect half-width (meters)}
}
\value{
Returns a list of 4 items: 1. "true_N" is a matrix of true abundance values for each site (row) and sample (column). 2. "n_obs" is a matrix of the number of observed organisms at each site (row) and sample (column). 3. "y_list" is a list of vectors. Each vector holds the distance data for a single survey. 4. "inputs" is a list of input values.
}
\description{
Generate point counts and/or distance sampling data that is biased by over-dispersion
in the true abundance, N. Distance sampling data is created with each dataset,
but it can just be ignored if you only want the point count data. This is a 
rudementary function. Currently it assumes constant density, detection probability,
transects length, etc.
}
\examples{
sim_data_N()

}
