% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_p.R
\name{sim_data_p}
\alias{sim_data_p}
\title{Simulate biased (via over-dispersion in p) point count and distance sampling data}
\usage{
sim_data_p(
  n_sites = 50,
  n_samps = 6,
  lambda = 10,
  mean_det_prob = 0.42,
  sigma_beta_dist_p = 0.01,
  alpha = NA,
  beta = NA,
  W = 20
)
}
\arguments{
\item{n_sites}{number of sites (transects)}

\item{n_samps}{number of samples (replicates) per site}

\item{lambda}{mean abundance at every site}

\item{mean_det_prob}{mean detection probability across all sites and replicates (different value drawn for each site and replicate)}

\item{sigma_beta_dist_p}{SD of a beta distribution from which the realized detection probability is drawn (keep this below ~ 0.16)}

\item{alpha}{parameter of the beta distribution. This is deterministically calculated from mean_det_prob and sigma_beta_dist_p if they are provided.}

\item{beta}{parameter of the beta distribution. This is deterministically calculated from mean_det_prob and sigma_beta_dist_p if they are provided.}

\item{W}{transect half-width (meters)}
}
\value{
Returns a list of 4 items: 1. "true_N" is a matrix of true abundance values for each site (row) and sample (column). 2. "n_obs" is a matrix of the number of observed organisms at each site (row) and sample (column). 3. "y_list" is a list of vectors. Each vector holds the distance data for a single survey. 4. "inputs" is a list of input values.
}
\description{
Generate point counts and/or distance sampling data that is biased by over-dispersion
in the detection probability, p. Distance sampling data is created with each dataset,
but it can just be ignored if you only want the point count data. This is a 
rudementary function. Currently it assumes constant density, detection probability,
transects length, etc.
}
\examples{
sim_data_p()

}
