% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function_N.R
\name{simulation_function_N}
\alias{simulation_function_N}
\title{Primary function for simulating & analysing datasets that are biased by over-dispersion in N (Scenario 2)}
\usage{
simulation_function_N(
  n_sites = 50,
  n_samps = 6,
  lambda1 = 5,
  lambda2 = 5,
  det_prob = 0.42,
  sigma = NA,
  W = 20,
  reps_to_analyze = 3,
  return = "results",
  savefilename = file.path("set 1", "datasets", "data")
)
}
\arguments{
\item{n_sites}{number of sites (transects)}

\item{n_samps}{number of samples (replicates) per site}

\item{lambda1}{partial mean abundance at every site (single draw per site that stays constant across samples)}

\item{lambda2}{partial mean abundance at every replicate (new draw for every sample at every site)}

\item{det_prob}{mean detection probability}

\item{sigma}{detection parameter (meters). Just leave this blank and it will be calculated from det_prob}

\item{W}{transect half-width (meters)}

\item{reps_to_analyze}{the number of samples/replicates to analyze. If NA, it will analyse all replicates in the data.}

\item{return}{What to return from the function call. Currently the only option is 'results'. May change this to only analyze simulated goodness-of-fit metrics.}

\item{savefilename}{The simulated datasets and results ARE saved to file (currently not optional). This provides the path and filename for saving the intermediate steps in the analysis.}
}
\value{
if everything works well, it returns a data.frame with the results of simulating a single dataset, analyzing it in 4 ways, and calculating randomized quantile residuals a la Knape et al. 2018. It also saves a list with the simulated dataset, dataframe of results (minus rqr residual info), and the actual rqr residuals to a savefilename inside the folder path 'working directory'/results/Scenario 2/savefilename. If there is an error, the function returns NA and also saves a file to 'working directory'/Scenario 2/set x/errors with the simulated dataset and the results data.frame but no rq-residuals (it's basically assumed that the rq-residuals were the source of the error.) Similarly, with a warning the function returns the results data.frame and also saves a file to 'working directory'/Scenario 2/set x/warnings with the simulated dataset and the results data.frame but no rq-residuals (it's basically assumed that the rq-residuals were the source of the error.) The user will have to go back and try to calculate rq-residuals from the output later.
}
\description{
This function is one of the master functions in this package. It simulates and
analyzes a single dataset that is over-dispersed in true abundance, N. 
To do this many times, use an apply function. It's currently simplified and 
assumes constant abundance, detection, transect length, NO simulated goodness-of-fit
metrics. It simulates both point count and distance data and analyses both datasets
using both unmarked and optim.
}
\examples{
simulation_function_N()

}
