% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function_nobs.R
\name{simulation_function_nobs}
\alias{simulation_function_nobs}
\title{Primary function for simulating & analysing datasets that are biased by over-dispersion in the number of observations (via double-counting) (Scenario 1)}
\usage{
simulation_function_nobs(
  n_sites = 50,
  n_samps = 6,
  lambda = 10,
  alpha = (1 - 0.42),
  gamma = 0.02,
  beta = NA,
  det_prob = NA,
  sigma = NA,
  W = 20,
  reps_to_analyze = 3,
  return = "results",
  savefilename = "set 1/datasets/data"
)
}
\arguments{
\item{n_sites}{number of sites (transects)}

\item{n_samps}{number of samples (replicates) per site}

\item{lambda}{mean abundance at every site (single draw per site that stays constant across samples)}

\item{alpha}{probability of NOT being observed (i.e. alpha = 1 - detection probability)}

\item{gamma}{probability of being observed exactly twice (gamma = 1 - alpha - beta)}

\item{beta}{(optional) probability of being observed exactly once. If not specified, it's calculated as: beta = 1 - alpha - gamma}

\item{det_prob}{(optional) probability of being observed. If not specified, it's calculated as: det_prob = beta + gamma}

\item{sigma}{(optional) detection parameter (meters). Best to leave this blank and it will be calculated from det_prob.}

\item{W}{transect half-width (meters)}

\item{reps_to_analyze}{the number of samples/replicates to analyze. If NA, it will analyse all replicates in the data.}

\item{return}{What to return from the function call. Currently the only option is 'results'. May change this to only analyze simulated goodness-of-fit metrics.}

\item{savefilename}{The simulated datasets and results ARE saved to file (currently not optional). This provides the path and filename for saving the intermediate steps in the analysis.}
}
\value{
if everything works well, it returns a data.frame with the results of simulating a single dataset, analyzing it in 4 ways, and calculating randomized quantile residuals a la Knape et al. 2018. It also saves a list with the simulated dataset, dataframe of results (minus rqr residual info), and the actual rqr residuals to a savefilename inside the folder path 'working directory'/results/Scenario 1/savefilename. If there is an error, the function returns NA and also saves a file to 'working directory'/Scenario 1/set x/errors with the simulated dataset and the results data.frame but no rq-residuals (it's basically assumed that the rq-residuals were the source of the error.) Similarly, with a warning the function returns the results data.frame and also saves a file to 'working directory'/Scenario 1/set x/warnings with the simulated dataset and the results data.frame but no rq-residuals (it's basically assumed that the rq-residuals were the source of the error.) The user will have to go back and try to calculate rq-residuals from the output later.
}
\description{
This function is one of the master functions in this package. It simulates and
analyzes a single dataset that is over-dispersed in observations, n_obs. 
To do this many times, use an apply function. It's currently simplified and 
assumes constant abundance, detection, and transect length. It does not (yet) 
simulate goodness-of-fit metrics. It simulates both point count and distance 
data and analyses both datasets using both unmarked and optim.
}
\examples{
simulation_function_N()

}
