% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/um_summary.R
\name{um_summary}
\alias{um_summary}
\title{Goodness-of-fit Statistics}
\usage{
um_summary(object)
}
\arguments{
\item{object}{the output from the fitstats() function}
}
\value{
data.frame with fit_obs = fit statistic of observed data; "mean(fit_obs - fit_sim)" = mean bias in the fit statistics; "StdDev(fit_obs - fit_sim)" = SD of the bias in fit statistics; "Pr(fit_sim > fit_obs)" = p-value of the fit statistics
}
\description{
I copied this function from https://github.com/rbchan/unmarked/blob/master/R/boot.R
It's a function returning a summary of the 3 fit statistics calculated using
fitstat().
}
